// Bullhorn-Attachments-UI-Code/src/app/shared/models/static-search-filters.model.ts
export interface StatusModel {
    name: string;
    id: number;
}

export interface RegisteredBUModel {
    name: string;
    id: number;
}

export interface ClientModel {
    name: string;
    id: number;
}

export interface ServiceLineModel {
    name: string;
    id: number;
}

export interface SupplierNameModel {
    name: string;
    id: number;
}

export interface ApproverModel {
    name: string;
    id: number;
}

export interface FinancialReviewModel{
    name:string;
    id:number;
}
export interface CESEESModel {
    name: string;
    id: number;
}