// Bullhorn-Attachments-UI-Code/src/app/shared/models/Data-Table/worker-search-data-grid/ws-data-tbl.model.ts
export interface WorkerSearchDataTable {
    caseId: number | null | undefined;
    atsPlacementId: number | null | undefined;
    workerName: string | null | undefined;
    prbuId: number | null | undefined;
    legacyRepotingBu: string | null | undefined;
    customerId: number | null | undefined;
    customerName: string | null | undefined;
    supplierId: string | null | undefined;
    supplierName: string | null | undefined;
    serviceLineId: number | null | undefined;
    serviceLine: string | null | undefined;
    branch: string | null | undefined;
    workerTypeId: number | null | undefined;
    workerType: string | null | undefined;
    cesEes: string | null | undefined;
    placementStatusId: number | null | undefined;
    placementReviewStatus: string | null | undefined;
    reviewedBy: string | null | undefined;
    reviewedDate: string | null | undefined;
    updatedBy: string | null | undefined;
    updatedDate: string | null | undefined;
    targetStartDate: string | null | undefined;
}